+++ bundles/org.palladiosimulator.dependability.reliability.uncertainty.solver/src/org/palladiosimulator/dependability/reliability/uncertainty/solver/jobs/UncertaintyBasedReliabilityPredictionJob.java
@@ -14,7 +14,7 @@ import org.palladiosimulator.analyzer.workflow.jobs.ValidatePCMModelsJob;
 import org.palladiosimulator.architecturaltemplates.jobs.RunATJob;
 import org.palladiosimulator.architecturaltemplates.jobs.config.ATExtensionConfigurationBuilder;
 import org.palladiosimulator.experimentautomation.application.jobs.PrepareBlackboardJob;
-import org.palladiosimulator.solver.runconfig.PCMSolverWorkflowRunConfiguration;
+import org.palladiosimulator.solver.core.runconfig.PCMSolverWorkflowRunConfiguration;
 
 import de.uka.ipd.sdq.workflow.jobs.ICompositeJob;
 import de.uka.ipd.sdq.workflow.jobs.SequentialBlackboardInteractingJob;
