+++ tests/org.palladiosimulator.dependability.reliability.uncertainty.solver.tests/test/org/palladiosimulator/dependability/reliability/uncertainty/solver/tests/ReliabilityPredictionTestDefinition.java
@@ -32,7 +32,7 @@ import org.palladiosimulator.experimentautomation.application.jobs.PrepareBlackb
 import org.palladiosimulator.reliability.solver.pcm2markov.MarkovTransformationResult;
 import org.palladiosimulator.reliability.solver.pcm2markov.Pcm2MarkovStrategy;
 import org.palladiosimulator.solver.core.models.PCMInstance;
-import org.palladiosimulator.solver.runconfig.PCMSolverWorkflowRunConfiguration;
+import org.palladiosimulator.solver.core.runconfig.PCMSolverWorkflowRunConfiguration;
 
 import com.google.common.collect.Lists;
